pyproject.toml
[project]
name = "hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "fpdf2>=2.8.4",
    "markdown>=3.8.2",
    "passlib[bcrypt]>=1.7.4",
    "portia-sdk-python>=0.7.2",
    "pydantic[email]>=2.11.7",
    "pygments>=2.19.2",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "reportlab>=4.4.3",
    "uvicorn[standard]>=0.35.0",
    "weasyprint>=66.0",
]
